Fix Python 3.10 issue with test_cmndata

+++ unittests/test_cmndata.py
@@ -50,7 +50,7 @@ class cmndata_tests(wtc.WidgetTestCase):
     def test_PD_PaperSize(self):
         pd = wx.PrintData()
         pd.GetPaperSize()
-        pd.SetPaperSize( wx.Size(8.5*300, 11*300) )
+        pd.SetPaperSize( wx.Size(int(8.5*300), 11*300) )
         pd.PaperSize
